% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteRobustLinearRegression.R
\name{ExecuteRobustLinearRegression}
\alias{ExecuteRobustLinearRegression}
\title{Execute RobustLinearRegression called with JSON from Cloudfoundry}
\usage{
ExecuteRobustLinearRegression(requestJSON)
}
\arguments{
\item{requestJSON}{the JSON stream passed on via http body when called with http POST}
}
\value{
responseJSON return JSON stream of http POST call
}
\description{
Execute RobustLinearRegression called with JSON from Cloudfoundry
}
\examples{
library(MASS);library(jsonlite);ExecuteRobustLinearRegression('{"x":[4, 4, 7, 7, 8, 9,10,10,10],"y":[2,10, 4,22,16,10,18,26,34]}')
}

