% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustLinearRegression.R
\name{RobustLinearRegression}
\alias{RobustLinearRegression}
\title{Fit a linear model by robust regression using an M estimator.}
\usage{
RobustLinearRegression(x, y, ...)
}
\arguments{
\item{x}{a matrix or data frame containing the explanatory variables.}

\item{y}{the response: a vector of length the number of rows of x.}

\item{...}{arguments passed on to MASS::rlm and underlying functions}
}
\value{
linear regression model
}
\description{
Fit a linear model by robust regression using an M estimator.
}
\examples{
library(MASS);RobustLinearRegression(cars$speed, cars$dist)
}

